--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -2,6 +2,7 @@
 package hw10programoptimization
 
 import (
+	"bufio"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -32,38 +33,50 @@ func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 	return countDomains(u, domain)
 }
 
-type users [100_000]User
+func getUsers(r io.Reader) ([]User, error) {
+	// content, err := io.ReadAll(r)
+	// if err != nil {
+	// 	return nil, err
+	// }
 
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := io.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
+	result := make([]User, 0)
+	scanner := bufio.NewScanner(r)
+	for scanner.Scan() {
+		line := scanner.Text()
 		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
-	result := make(DomainStat)
-
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
+		if err := json.Unmarshal([]byte(line), &user); err != nil {
 			return nil, err
 		}
+		result = append(result, user)
+	}
+
+	// lines := strings.Split(string(content), "\n")
+	// result := make([]User, len(lines))
+	// for i, line := range lines {
+	// 	var user User
+	// 	if err = json.Unmarshal([]byte(line), &user); err != nil {
+	// 		return nil, err
+	// 	}
+	// 	result[i] = user
+	// }
+	return result, nil
+}
+
+func countDomains(u []User, domain string) (DomainStat, error) {
+	result := make(DomainStat)
+	expression, err := regexp.Compile("\\." + domain)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, user := range u {
+		matched := expression.Match([]byte(user.Email))
 
 		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
+			domain := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+			num := result[domain]
 			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+			result[domain] = num
 		}
 	}
 	return result, nil