--- a/hw10_program_optimization/stats_optimization_test.go
+++ b/hw10_program_optimization/stats_optimization_test.go
@@ -1,6 +1,7 @@
 //go:build bench
 // +build bench
 
+//nolint:godot
 package hw10programoptimization
 
 import (