## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [x] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [x] Добавлены юнит-тесты - до 3 баллов
- [x] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

__Итоговое сравнение результатов бенчмарка:__
```
goos: linux
goarch: amd64
pkg: github.com/Averlex/golang-hw/hw10_program_optimization
cpu: AMD Ryzen AI 9 HX 370 w/ Radeon 890M
                 │ bench_old.txt │            bench_new.txt            │
                 │    sec/op     │   sec/op     vs base                │
GetDomainStat-24    362.48m ± 3%   61.44m ± 4%  -83.05% (p=0.000 n=10)

                 │ bench_old.txt │            bench_new.txt             │
                 │     B/op      │     B/op      vs base                │
GetDomainStat-24   321.96Mi ± 0%   24.00Mi ± 0%  -92.54% (p=0.000 n=10)

                 │ bench_old.txt │            bench_new.txt            │
                 │   allocs/op   │  allocs/op   vs base                │
GetDomainStat-24    2844.4k ± 0%   822.2k ± 0%  -71.09% (p=0.000 n=10)
```

__Результаты предоставленной задании функции с бенчем ДО изменений:__
stats_optimization_test.go:46: time used: 437.082553ms / 300ms
stats_optimization_test.go:47: memory used: 328Mb / 30Mb

__Результаты предоставленной задании функции с бенчем ПОСЛЕ изменений:__
stats_optimization_test.go:51: time used: 156.451063ms / 300ms
stats_optimization_test.go:52: memory used: 24Mb / 30Mb

__CPU профиль до изменений:__
<img width="1510" height="889" alt="cpu_profile_before" src="https://github.com/user-attachments/assets/0dd360c1-8125-4580-8d7c-6a5afd8bf8af" />

__Memory профиль до изменений:__
<img width="875" height="887" alt="memory_profile_before" src="https://github.com/user-attachments/assets/4f433ecf-624c-49f8-a0e2-4236c63d9b03" />

__CPU профиль после изменений:__
<img width="1214" height="810" alt="cpu_profile_after" src="https://github.com/user-attachments/assets/f9921165-73a0-4e58-87e0-9aa83e1aa8a1" />

__Memory профиль после изменении:__
Сюда почему-то попадает память, аллоцируемая в тесте до профилирования. Она собой, вероятно, перекрывает память, используемую функцией, но результат всё равно видно.
<img width="684" height="691" alt="memory_profile_after" src="https://github.com/user-attachments/assets/d3363cbc-316f-4c75-914a-c1b0f3677f6f" />

__Порядок применения оптимизаций:__
1. regexp.Match заменен на Compile вне цикла + Match внутри
2. Массив users создаётся динамически - от количество строк в файле -> меньше памяти
3. Чтение из Reader в []byte с последующим сплитом на строки фактически дублировало данные -> заменено на чтение из Reader напрямую
4. Анмаршаллинг изменен на easyjson -> меньше памяти и времени
5. Сканирование сразу в байты, а не в строки -> меньше памяти и немного меньше времени на преобразования

К этому моменту результаты бенча выглядели следующим образом:
stats_optimization_test.go:51: time used: 178.260954ms / 300ms
stats_optimization_test.go:52: memory used: 64Mb / 30Mb

6. Изменена логика хэлперов - теперь обработка идёт построчно, промежуточные результаты не накапливаются, только конечная мапа -> меньше памяти